+++ products/editor/backend/chunking/__init__.py
@@ -1,3 +1,5 @@
+from typing import cast
+
 from .code import chunk_code as _chunk_code
 from .exceptions import UnsupportedLanguage
 from .parser import ProgrammingLanguage
@@ -6,8 +8,8 @@ from .text import chunk_text as _chunk_text
 
 def chunk_text(language: str, content: str, chunk_size: int = 300, chunk_overlap: float = 0.2):
     try:
-        if language in ProgrammingLanguage:
-            return _chunk_code(language, content, chunk_size, chunk_overlap)
+        if language in {member.value for member in ProgrammingLanguage}:
+            return _chunk_code(cast(ProgrammingLanguage, language), content, chunk_size, chunk_overlap)
     except UnsupportedLanguage:
         pass
     return _chunk_text(content, chunk_size, chunk_overlap)
